V1.1 UPDATE
Basically The final version
May come back to this, but really making it in pygame and not using and guides means that there is proabably big design challenges that need adressing. While I think it would be super cool to address these, and make an awesome Tetris clone, I feel that the effort would be wasted, since I have other skills that need developing e.g. web scraping, SQL, general C++ (for half life mods lol) etc.

This final version has the New Piece Que and the only issues in terms of gameplay I still have is:
pygame keyboard event handling - you can spam alot of things at once, or accidently rotate too much
No menus or options - When/if i decide to port this to KIVY, I will adress this, but because its in pygame it just sucks alot